Use `toBase` instead of `.toString()` to get the room index
Getting a full book's info on Chrome on Android took about 3 minutes.

It turned out that it was the `.toString()` call to get the room index.

V8's source code shows that the fast .toString algorithm is behind the flag `V8_ADVANCED_BIGINT_ALGORITHMS`.

I guess this flag is ON when Chrome is built for Linux desktop, but OFF when it's built for Android.

Replacing this `.toString()` by a call to `toBase(...)` fixed this issue.